moved hamburger button component to its own folder, fixed typo in hamburger button id
