+++ src/components/HamburgerButton/HambugerButton.js
@@ -3,7 +3,7 @@ import './HamburgerButton.css'
 
 const HamburgerButton = (props) => {
     return(
-        <div onClick = {props.onClick} className={`hamburger ${props.activated ? 'is-active' : ''}`} id="hamburger-6">
+        <div onClick = {props.onClick} className={`hamburger ${props.activated ? 'is-active' : ''}`} id="hamburger">
           <span className="line"></span>
           <span className="line"></span>
           <span className="line"></span>
