+++ src/components/HamburgerButton/HamburgerButton.css
@@ -31,7 +31,7 @@ html {
     HAMBURGER BUTTON ANIMATIONS
 --------------------------------------
 */
-#hamburger-6.is-active{
+#hamburger.is-active{
     -webkit-transition: all 0.3s ease-in-out;
     -o-transition: all 0.3s ease-in-out;
     transition: all 0.3s ease-in-out;
@@ -44,25 +44,25 @@ html {
     transform: rotate(45deg);
   }
   
-  #hamburger-6.is-active .line:nth-child(2){
+  #hamburger.is-active .line:nth-child(2){
     width: 0px;
   }
   
-  #hamburger-6.is-active .line:nth-child(1),
-  #hamburger-6.is-active .line:nth-child(3){
+  #hamburger.is-active .line:nth-child(1),
+  #hamburger.is-active .line:nth-child(3){
     -webkit-transition-delay: 0.3s;
     -o-transition-delay: 0.3s;
     transition-delay: 0.3s;
   }
   
-  #hamburger-6.is-active .line:nth-child(1){
+  #hamburger.is-active .line:nth-child(1){
     -webkit-transform: translateY(13px);
     -ms-transform: translateY(13px);
     -o-transform: translateY(13px);
     transform: translateY(13px);
   }
   
-  #hamburger-6.is-active .line:nth-child(3){
+  #hamburger.is-active .line:nth-child(3){
     -webkit-transform: translateY(-13px) rotate(90deg);
     -ms-transform: translateY(-13px) rotate(90deg);
     -o-transform: translateY(-13px) rotate(90deg);
